Specify fsid-secret for initial monitor node
node['ceph']['fsid-secret'] only applies if ceph_chef_mon_nodes is not empty
and the 0th monitor returned by a chef search does not contain an fsid-secret attribute.

This update makes it so the initial monitor can pull from node['ceph']['fsid-secret']
if it exists, otherwise the usual behavior of generating an fsid-secret will apply.